plusieurs problèmes réglés concernant le front + raccourci des méthodes + changé la détection des clones + changé tout l'ASM

+++ src/main/java/fr/uge/clone/model/CloneArtefact.java
@@ -1,9 +1,4 @@
 package fr.uge.clone.model;
 
-/**
- * Record representing a Clone Artefact.
- * @param artefact the artefact
- * @param score the score of the clone
- */
-public record CloneArtefact(Artefact artefact, long score) {
+public record CloneArtefact(Artefact artefact, double score) {
 }
